- Refactored notification into the ViewModel - Fixed notification being dismissed automatically
