+++ app/src/main/java/org/british_information_technologies/gym_log_book/activity/exercise_set_guide_activity/Main.kt
@@ -1,7 +1,11 @@
 package org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity
 
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.*
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
@@ -9,6 +13,11 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import org.british_information_technologies.gym_log_book.R
+import org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.page.AskExtraSetPage
+import org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.page.ExerciseSetGuideActivityPage
+import org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.page.PausePage
+import org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.page.SetPage
+import org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.page.StartPage
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -52,5 +61,4 @@ fun Main(
 			}
 		}
 	)
-
 }
