+++ app/src/main/java/org/british_information_technologies/gym_log_book/activity/exercise_set_guide_activity/page/StartPage.kt
@@ -1,4 +1,4 @@
-package org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity
+package org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.page
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -14,6 +14,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import org.british_information_technologies.gym_log_book.activity.amend_exercise_activity_v2.components.ExerciseTypeDropdownSelector
+import org.british_information_technologies.gym_log_book.activity.exercise_set_guide_activity.SetGuideViewModelV2
 
 @Composable
 fun StartPage(
