+++ app/src/main/res/values/themes.xml
@@ -1,13 +1,16 @@
 <resources>
 	<!-- Base application theme. -->
 	<style name="Theme.Gym_Log_Book" parent="Theme.Material3.DynamicColors.DayNight">
-		<!-- Status bar color. -->
-		<item name="android:windowIsTranslucent">true</item>
 		<item name="android:windowBackground">@android:color/transparent</item>
-		<item name="android:windowContentOverlay">@null</item>
-		<item name="android:windowNoTitle">true</item>
-		<item name="android:windowIsFloating">false</item>
 		<item name="android:backgroundDimEnabled">true</item>
+		<item name="windowActionBar">false</item>
+		<item name="windowNoTitle">true</item>
+	</style>
+
+	<style name="Theme.Gym_Log_Book.Dialogue" parent="Theme.Material3.DynamicColors.DayNight">
+		<item name="android:windowBackground">@android:color/transparent</item>
+		<item name="android:backgroundDimEnabled">true</item>
+		<item name="android:windowIsTranslucent">true</item>
 		<item name="windowActionBar">false</item>
 		<item name="windowNoTitle">true</item>
 	</style>
